Quick add task bug fixed, needed to convert cookies.id and project_id  to number before database
